pyproject.toml
[tool.poetry]
name = "countdowner"
version = "3.1.0"
description = "A library to check for sales at Countdown grocery stores throughout New Zealand."
authors = ["Alex Raichev <alex@raichev.net>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7||^3.8"
requests = "^2.23.0"
pandas = "^1.0.4"
click = "^7.1.2"
pyyaml = "^5.3.1"
lxml = "^4.5.1"
yagmail = "^0.14"
keyring = "^21.2.1"
titlecase = "^1.1.1"

[tool.poetry.dev-dependencies]
jupyter = "^1.0.0"
pytest = "^5.4.3"
responses = "^0.10.14"
python-dotenv = "^0.13.0"
pre-commit = "^2.5.1"
black = "^19.10b0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
